fix: compatibility with nixos-install
`nixos-install` copies the config and all flake inputs to `/mnt/nix/store` by default, but when run from within the installer, the `outPath` of the input derivations still points to their original location within `/nix/store`, not `/mnt/nix/store`.
The nixpkgs documentation module relies on `modulesPath` having the same prefix as its file location, which is not the case here.
This commit provides a workaround for this problem by not relying on the `outPath` of flake inputs for the `modulesPath` os arg.